Refactor into an internal _samplers sub-package.
