+++ src/somd2/runner/_repex.py
@@ -856,7 +856,7 @@ class RepexRunner(_RunnerBase):
 
         # Create the terminal flip sampler (if terminal groups were detected).
         if self._terminal_groups:
-            from ._terminal_flip import TerminalFlipSampler
+            from ._samplers import TerminalFlipSampler
 
             self._terminal_flip_sampler = TerminalFlipSampler(
                 self._terminal_groups,
